LPS-56481 Rename package
