+++ modules/apps/message-boards/message-boards-comment-service/src/com/liferay/message/boards/comment/service/MBDiscussionImpl.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.message.boards.comment.impl;
+package com.liferay.message.boards.comment.service;
 
 import com.liferay.portal.kernel.comment.Discussion;
 import com.liferay.portal.kernel.comment.DiscussionComment;
